+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 #### New Features
 - Added partial support for `Series.str.translate` where the values in the `table` are single-codepoint strings.
 - Added support for `DataFrame.corr`.
+- Added support for `DataFrame.compare` and `Series.compare` with default parameters.
 
 ## 1.19.0 (2024-06-25)